Refactor shared skiptracer logic into helper module
